refactor: accidentally the exceptions folder for file was placed in the root directory
