+++ backend/apps/innogram/src/file/file.service.ts
@@ -6,7 +6,7 @@ import { FileOutboxRepository } from './file-outbox.repository';
 import { FileAction, type File } from '@prisma/client';
 import { type FileOutboxWithFile } from './utils/file-outbox-with-file-select.util';
 import { Transactional } from '@nestjs-cls/transactional';
-import { FileOutboxNotFoundError } from '../exceptions';
+import { FileOutboxNotFoundError } from './exceptions';
 
 @Injectable()
 export class FileService {
